Paddle position is statefull now
Now the paddle position does not reset when releasing the joystick.
Limited the paddle position within the TFT screen.
Enabled debug output: It prints the paddle location when clicking the
joystick.

Todo:
Need to update the circuit schema
Known bug: The paddle is drawn as 5px tall but it behaves as 1px tall
when interacting with the ball.